requirements.txt
absl-py==1.1.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==5.0.0
CacheControl==0.12.11
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==7.1.2
colorama==0.4.4
compact==0.1.2
comtypes==1.1.10
contextlib2==21.6.0
covid==2.4.0
cycler==0.10.0
Cython==0.29.30
DateTime==4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
Django==4.0
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djitellopy==2.3.1
docopt==0.6.2
entrypoints==0.4
enum34==1.1.10
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.12
firebase-admin==5.2.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
gcloud==0.17.0
google==3.0.0
google-api-core==2.4.0
google-api-python-client==2.45.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.54.0
googlesearch-python==1.1.0
greenlet==1.1.1
grpcio==1.44.0
grpcio-status==1.44.0
gTTS==2.2.2
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httpcore==0.14.7
httplib2==0.20.2
httpx==0.22.0
idna==2.10
iexfinance==0.5.0
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
iso8601==1.0.2
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.0.1
Js2Py==0.70
jsonschema==4.6.0
jupyter-client==7.3.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jws==0.1.3
kaggle==1.5.12
keras==2.9.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelImg==1.8.6
libclang==14.0.1
lxml==4.9.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mediapipe==0.8.6
mistune==0.8.4
MouseInfo==0.1.3
msgpack==1.0.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
node==0.9.26
notebook==6.4.12
numpy==1.20.2
oauth2client==3.0.0
oauthlib==3.1.1
odict==1.7.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.8
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.1.2
pipenv==2021.11.23
pipwin==0.5.0
plumber==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.8.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycaw==20181226
pycparser==2.21
pycryptodome==3.4.3
pydantic==1.8.1
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.12.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyPrind==2.11.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
Pyrebase==3.0.27
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.26
pyserial==3.5
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jwt==2.0.1
python-slugify==6.1.1
PythonTurtle==0.3.2
pytube==10.5.3
pytube3==9.6.4
PyTweening==1.0.3
pytz==2021.3
pywin32==304
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==23.1.0
QtPy==2.1.0
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rfc3986==1.5.0
rsa==4.7.2
scipy==1.7.0
Send2Trash==1.8.0
serial==0.0.97
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.23
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
threadpoolctl==2.1.0
tinycss2==1.1.1
torch==1.11.0
tornado==6.1
tqdm==4.63.0
traitlets==5.2.2.post1
typer==0.3.2
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.2
virtualenv==20.4.7
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
youtube-search-python==1.6.5
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
